Fixed an major soundness issue, and bug in Match.do_infer().
There was a bug in Match.do_infer() where it failed to apply the final substitution of the matchand for as_term in the return type.
Separately from this, compare_terms() had a huge bug where it was indifferent to the alpha-renaming of variables bound in the context.
This was a massive hole that clearly yielded unsoundness, and would admit ill-typed terms all over the place.